init - move home to it's own dir
