+++ weave-js/src/components/PagePanelComponents/Home/Home.tsx
@@ -1,16 +1,18 @@
 import {Node, voidNode} from '@wandb/weave/core';
 import moment from 'moment';
 import React, {FC, memo, useCallback, useState} from 'react';
-import getConfig from '../../config';
+import getConfig from '../../../config';
 
 import styled from 'styled-components';
-import {WBButton} from '../../common/components/elements/WBButtonNew';
-import {useWeaveContext} from '../../context';
-import {ChildPanelFullConfig} from '../Panel2/ChildPanel';
-import {IconAddNew as IconAddNewUnstyled, IconWeaveLogo} from '../Panel2/Icons';
-import {PanelRootBrowser} from '../Panel2/PanelRootBrowser/PanelRootBrowser';
-import {useNewPanelFromRootQueryCallback} from '../Panel2/PanelRootBrowser/util';
-import {dummyProps, useConfig} from '../Panel2/panel';
+import {WBButton} from '../../../common/components/elements/WBButtonNew';
+import {useWeaveContext} from '../../../context';
+import {ChildPanelFullConfig} from '../../Panel2/ChildPanel';
+import {
+  IconAddNew as IconAddNewUnstyled,
+  IconWeaveLogo,
+} from '../../Panel2/Icons';
+import {useNewPanelFromRootQueryCallback} from '../../Panel2/PanelRootBrowser/util';
+import {useConfig} from '../../Panel2/panel';
 
 type HomeProps = {
   updateConfig: (newConfig: ChildPanelFullConfig) => void;
@@ -74,15 +76,7 @@ const HomeComp: FC<HomeProps> = props => {
         </TopBarRight>
       </TopBar>
       <BrowserContainer>
-        <PanelRootBrowser
-          input={voidNode() as any}
-          updateInput={updateInput as any}
-          isRoot={true}
-          config={rootConfig}
-          updateConfig={updateRootConfig}
-          context={dummyProps.context}
-          updateContext={dummyProps.updateContext}
-        />
+        <>Hi</>
       </BrowserContainer>
     </>
   );
